batik-classifier/api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0
gunicorn==21.2.0
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-multipart==0.0.9
